Remove testing-related libraries and associated code
This removes testing-related dependencies and action, reducer, selector, saga, and type files for cart, user, and categories. The testing setup, including Jest, DOM Matchers, and testing user and category sagas, and the redux store setup for testing are also deleted. This decision is based on a shift in testing strategy and requirements. The cleanup addresses redundancy and improves maintainability. It's important to update the testing and continuous integration pipelines accordingly. This change addresses internal refactoring and does not affect end-users directly.

No relevant issues apply.