Refactor dependencies and add CaptchaProvider component
This change includes adding new dependencies, such as "@hookform/resolvers" and "@types/react-router-dom", for form handling and routing, and creating a new CaptchaProvider component to manage the captcha functionality within the application. The CaptchaProvider utilizes the HCaptcha library and provides context for rendering captchas and handling verification. This commits also refactors the code for button components and online forms to integrate with the new CaptchaProvider. Overall, the changes aim to enhance security and user experience in form submissions.